+++ cpp/CHANGELOG.md
@@ -21,6 +21,10 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-92 - Unify behavior of debug and release build
 - MS-98 - Install all unit test to installation directory
 - MS-115 - Change is_startup of metric_config switch from true to on
+- MS-122 - Archive criteria config 
+- MS-124 - HasTable interface
+- MS-126 - Add more error code
+
 ## New Feature
 
 - MS-57 - Implement index load/search pipeline